feat(lua): add DB set/get for custom user data
Add db_get and db_set methods to to M module to enable
storing custom user data in memory or permanent storage.

Permanent storage flushes the user dato to the mink sqlite db,
while the in-mem storage is only temporary and its scope is
limited to the lifetime of the "sysagent" process.