fix(umlua): allow SIGNAL handler to return ""
Allow Signal handler to return with success when the
underlying Lua script returns an empty Lua string("").

Consider this case a non-error operation since the result
set comprises two values: string size and a value pointer.
When the size output value is zero, the value pointer
CAN be a NULL pointer.